Blockshuffle encryption
An experimental approach of how to extend or create a new cipher from
scratch. Taking an encrypted block and rearranging and encrypting to
cover a much larger area of 576 bits. The 9th block can be reserved for
salt rather than encrypting the data.

Also as an experiment a rng based on the 576bit encryption will cycle
near infinitely with random data even with simple seeds.

note: This hasn't been checked for obvious holes, the RNG in question
may be weak to start with, and the seeds haven't been checked to ensure
no two overlap. few verification checks are done, instead outputting to
give you a visual output of the results.